chore: improve code readability
