+++ contrib/core-contract-tests/tests/clar/utils/test-helpers.ts
@@ -2,6 +2,12 @@ import { ParsedTransactionResult } from "@hirosystems/clarinet-sdk";
 import { Cl, ClarityType, cvToString } from "@stacks/transactions";
 import { expect } from "vitest";
 
+/**
+ * checks whether the function is a valid test function starting with "test-"
+ * and not having any arguments
+ * @param functionCall
+ * @returns
+ */
 export function isValidTestFunction(functionCall: any) {
   if (functionCall.name.startsWith("test-") && functionCall.args.length > 0) {
     throw new Error("test functions must not have arguments");
@@ -11,6 +17,13 @@ export function isValidTestFunction(functionCall: any) {
   );
 }
 
+/**
+ * expect the result of tx index in the given block to be (ok true)
+ * @param block 
+ * @param contractFQN 
+ * @param functionName 
+ * @param index 
+ */
 export function expectOkTrue(
   block: ParsedTransactionResult[],
   contractFQN: string,
@@ -26,6 +39,13 @@ export function expectOkTrue(
   ).toBeOk(Cl.bool(true));
 }
 
+/**
+ * expect the result of tx index in the given block to be succesfull (ok ...)
+ * @param block 
+ * @param contractFQN 
+ * @param functionName 
+ * @param index 
+ */
 export function expectOk(
   block: ParsedTransactionResult[],
   contractFQN: string,
